+++ src/Proxy.ts
@@ -1,4 +1,5 @@
 import { QlikRepoApi } from "./main";
+import { UpdateCommonProperties } from "./util/UpdateCommonProps";
 
 import {
   IHttpStatus,
@@ -12,7 +13,11 @@ import {
   IServerNodeConfigurationCondensed,
 } from "./interfaces";
 
-import { IVirtualProxyUpdate } from "./interfaces/argument.interface";
+import {
+  IVirtualProxyUpdate,
+  IProxyCreate,
+  IProxyUpdate,
+} from "./interfaces/argument.interface";
 
 export class Proxy {
   constructor() {}
@@ -85,12 +90,124 @@ export class Proxy {
       .then((m) => m.data as Buffer);
   }
 
-  // TODO: complete implementation
+  // TODO: handle oidc arguments
+  public async proxyCreate(
+    this: QlikRepoApi,
+    arg: IProxyCreate
+  ): Promise<IProxyService> {
+    if (!arg.description)
+      throw new Error(`proxyCreate: "description" parameter is required`);
+    if (!arg.sessionCookieHeaderName)
+      throw new Error(
+        `proxyCreate: "sessionCookieHeaderName" parameter is required`
+      );
+
+    let data = {
+      description: arg.description,
+      sessionCookieHeaderName: arg.sessionCookieHeaderName,
+    };
+
+    if (arg.loadBalancingServerNodes) {
+      data["loadBalancingServerNodes"] = await parseLoadBalancingNodes(
+        this,
+        arg.loadBalancingServerNodes
+      );
+    }
+    if (arg.authenticationMethod)
+      data["authenticationMethod"] = parseAuthenticationMethod(
+        arg.authenticationMethod
+      );
+    if (arg.prefix) data["prefix"] = arg.prefix;
+    if (arg.description) data["description"] = arg.description;
+    if (arg.authenticationModuleRedirectUri)
+      data["authenticationModuleRedirectUri"] =
+        arg.authenticationModuleRedirectUri;
+    if (arg.websocketCrossOriginWhiteList)
+      data["websocketCrossOriginWhiteList"] = arg.websocketCrossOriginWhiteList;
+    if (arg.additionalResponseHeaders)
+      data["additionalResponseHeaders"] = arg.additionalResponseHeaders;
+    if (arg.sessionInactivityTimeout)
+      data["sessionInactivityTimeout"] = arg.sessionInactivityTimeout;
+
+    if (arg.samlMetadataIdP) data["samlMetadataIdP"] = arg.samlMetadataIdP;
+    if (arg.samlHostUri) data["samlHostUri"] = arg.samlHostUri;
+    if (arg.samlEntityId) data["samlEntityId"] = arg.samlEntityId;
+    if (arg.samlAttributeUserId)
+      data["samlAttributeUserId"] = arg.samlAttributeUserId;
+    if (arg.samlAttributeUserDirectory)
+      data["samlAttributeUserDirectory"] = arg.samlAttributeUserDirectory;
+    if (arg.samlAttributeMap) {
+      data["samlAttributeMap"] = parseSamlAttributeMap(arg.samlAttributeMap);
+    }
+    if (arg.samlSlo) data["samlSlo"] = arg.samlSlo;
+    if (arg.jwtPublicKeyCertificate)
+      data["jwtPublicKeyCertificate"] = arg.jwtPublicKeyCertificate;
+    if (arg.jwtAttributeUserId)
+      data["jwtAttributeUserId"] = arg.jwtAttributeUserId;
+    if (arg.jwtAttributeUserDirectory)
+      data["jwtAttributeUserDirectory"] = arg.jwtAttributeUserDirectory;
+    if (arg.jwtAttributeMap) {
+      data["jwtAttributeMap"] = parseJwtAttributeMap(arg.jwtAttributeMap);
+    }
+
+    return await this.repoClient
+      .Post(`virtualproxyconfig`, { ...data })
+      .then((res) => {
+        return res.data as IProxyService;
+      });
+  }
+
+  public async proxyUpdate(
+    this: QlikRepoApi,
+    arg: IProxyUpdate
+  ): Promise<IProxyService> {
+    if (!arg.id) throw new Error(`proxyUpdate: "id" parameter is required`);
+
+    validateRanges(arg);
+
+    let proxy = await this.proxyGet(arg.id).then((v) => v[0] as IProxyService);
+
+    if (arg.listenPort) proxy.settings.listenPort = arg.listenPort;
+    if (arg.unencryptedListenPort)
+      proxy.settings.unencryptedListenPort = arg.unencryptedListenPort;
+    if (arg.authenticationListenPort)
+      proxy.settings.authenticationListenPort = arg.authenticationListenPort;
+    if (arg.unencryptedAuthenticationListenPort)
+      proxy.settings.unencryptedAuthenticationListenPort =
+        arg.unencryptedAuthenticationListenPort;
+    if (arg.keepAliveTimeoutSeconds)
+      proxy.settings.keepAliveTimeoutSeconds = arg.keepAliveTimeoutSeconds;
+    if (arg.maxHeaderSizeBytes)
+      proxy.settings.maxHeaderSizeBytes = arg.maxHeaderSizeBytes;
+    if (arg.maxHeaderLines) proxy.settings.maxHeaderLines = arg.maxHeaderLines;
+    if (arg.restListenPort) proxy.settings.restListenPort = arg.restListenPort;
+    if (arg.allowHttp) proxy.settings.allowHttp = arg.allowHttp;
+    if (arg.sslBrowserCertificateThumbprint)
+      proxy.settings.sslBrowserCertificateThumbprint =
+        arg.sslBrowserCertificateThumbprint;
+    if (arg.kerberosAuthentication)
+      proxy.settings.kerberosAuthentication = arg.kerberosAuthentication;
+    if (!arg.virtualProxies) proxy.settings.virtualProxies = [];
+    if (arg.virtualProxies)
+      proxy.settings.virtualProxies = await parseVirtualProxies(
+        this,
+        arg.virtualProxies
+      );
+
+    let updateCommon = new UpdateCommonProperties(this, proxy, arg);
+    proxy = await updateCommon.updateAll();
+
+    return await this.repoClient
+      .Post(`proxyservice/${arg.id}`, { ...proxy })
+      .then((res) => res.data as IProxyService);
+  }
+
+  // TODO: what is this method supposed to do?
   public async virtualProxyAdd(
     this: QlikRepoApi,
-    virtualProxyId: string,
-    loadBalancingServerNodes?: string[],
-    websocketCrossOriginWhiteList?: string[]
+    virtualProxyId: string
+    // loadBalancingServerNodes?: string[],
+    // websocketCrossOriginWhiteList?: string[]
   ): Promise<IVirtualProxyConfig> {
     if (!virtualProxyId)
       throw new Error(`virtualProxyAdd: "virtualProxyId" is required`);
@@ -99,14 +216,6 @@ export class Proxy {
       (vp) => vp[0] as IVirtualProxyConfig
     );
 
-    // if (loadBalancingServerNodes)
-    //   virtualProxy.loadBalancingServerNodes += loadBalancingServerNodes;
-
-    // proxy.settings.virtualProxies.push(virtualProxy);
-
-    // return await this.repoClient
-    //   .Post(`proxyservice/${proxyId}`, proxy)
-    //   .then((res) => res.data as IVirtualProxyConfig);
     return virtualProxy;
   }
 
@@ -187,16 +296,9 @@ export class Proxy {
     if (arg.magicLinkFriendlyName)
       virtualProxy.magicLinkFriendlyName = arg.magicLinkFriendlyName;
     if (arg.authenticationMethod) {
-      if (arg.authenticationMethod == "Ticket")
-        virtualProxy.authenticationMethod = 0;
-      if (arg.authenticationMethod == "static")
-        virtualProxy.authenticationMethod = 1;
-      if (arg.authenticationMethod == "dynamic")
-        virtualProxy.authenticationMethod = 2;
-      if (arg.authenticationMethod == "SAML")
-        virtualProxy.authenticationMethod = 3;
-      if (arg.authenticationMethod == "JWT")
-        virtualProxy.authenticationMethod = 4;
+      virtualProxy.authenticationMethod = parseAuthenticationMethod(
+        arg.authenticationMethod
+      );
     }
     if (arg.samlMetadataIdP) virtualProxy.samlMetadataIdP = arg.samlMetadataIdP;
     if (arg.samlHostUri) virtualProxy.samlHostUri = arg.samlHostUri;
@@ -268,4 +370,61 @@ async function parseLoadBalancingNodes(
 
     return nodeCondensed[0];
   });
+}
+
+function parseAuthenticationMethod(authenticationMethod: string): number {
+  if (authenticationMethod == "Ticket") return 0;
+  if (authenticationMethod == "static") return 1;
+  if (authenticationMethod == "dynamic") return 2;
+  if (authenticationMethod == "SAML") return 3;
+  if (authenticationMethod == "JWT") return 4;
+
+  return 0;
+}
+
+function validateRanges(arg: IProxyUpdate) {
+  if (arg.listenPort < 1 && arg.listenPort > 65536)
+    throw new Error(`proxyUpdate: "listerPort" must be between 1 and 655536`);
+  if (arg.unencryptedListenPort < 1 && arg.unencryptedListenPort > 65536)
+    throw new Error(
+      `proxyUpdate: "unencryptedListenPort" must be between 1 and 655536`
+    );
+  if (arg.authenticationListenPort < 1 && arg.authenticationListenPort > 65536)
+    throw new Error(
+      `proxyUpdate: "authenticationListenPort" must be between 1 and 655536`
+    );
+  if (
+    arg.unencryptedAuthenticationListenPort < 1 &&
+    arg.unencryptedAuthenticationListenPort > 65536
+  )
+    throw new Error(
+      `proxyUpdate: "unencryptedAuthenticationListenPort" must be between 1 and 655536`
+    );
+  if (arg.keepAliveTimeoutSeconds < 1 && arg.keepAliveTimeoutSeconds > 300)
+    throw new Error(
+      `proxyUpdate: "keepAliveTimeoutSeconds" must be between 1 and 300`
+    );
+  if (arg.maxHeaderSizeBytes < 512 && arg.maxHeaderSizeBytes > 131072)
+    throw new Error(
+      `proxyUpdate: "maxHeaderSizeBytes" must be between 512 and 131072`
+    );
+  if (arg.maxHeaderLines < 20 && arg.maxHeaderLines > 1000)
+    throw new Error(
+      `proxyUpdate: "maxHeaderLines" must be between 20 and 1000`
+    );
+  if (arg.restListenPort < 1 && arg.restListenPort > 65536)
+    throw new Error(
+      `proxyUpdate: "restListenPort" must be between 1 and 655536`
+    );
+}
+
+async function parseVirtualProxies(
+  repoApi: QlikRepoApi,
+  vpArg: string[]
+): Promise<IVirtualProxyConfigCondensed[]> {
+  let allVP = await repoApi.virtualProxyGet();
+  let vpToAdd = allVP.filter((v) => {
+    return vpArg.includes(v.prefix);
+  });
+  return vpToAdd;
 }