+++ src/interfaces/argument.interface.ts
@@ -426,3 +426,52 @@ export interface ISharedContentCreate {
   type?: string;
   description?: string;
 }
+
+export interface IProxyCreate {
+  prefix?: string;
+  description: string;
+  sessionCookieHeaderName: string;
+  authenticationModuleRedirectUri?: string;
+  loadBalancingServerNodes?: string[];
+  websocketCrossOriginWhiteList?: string;
+  additionalResponseHeaders?: string;
+  authenticationMethod?:
+    | "Ticket"
+    | "HeaderStaticUserDirectory"
+    | "HeaderDynamicUserDirectory"
+    | "static"
+    | "dynamic"
+    | "SAML"
+    | "JWT";
+  samlMetadataIdP?: string;
+  samlHostUri?: string;
+  samlEntityId?: string;
+  samlAttributeUserId?: string;
+  samlAttributeUserDirectory?: string;
+  samlAttributeMap?: string[];
+  samlSlo?: boolean;
+  // samlSigningAlgorithm?: "sha1" | "sha256",
+  jwtPublicKeyCertificate?: string;
+  jwtAttributeUserId?: string;
+  jwtAttributeUserDirectory?: string;
+  jwtAttributeMap?: string[];
+  sessionInactivityTimeout?: number;
+}
+
+export interface IProxyUpdate {
+  id: string;
+  listenPort?: number;
+  allowHttp?: boolean;
+  unencryptedListenPort?: number;
+  authenticationListenPort?: number;
+  kerberosAuthentication?: boolean;
+  unencryptedAuthenticationListenPort?: number;
+  sslBrowserCertificateThumbprint?: string;
+  keepAliveTimeoutSeconds?: number;
+  maxHeaderSizeBytes?: number;
+  maxHeaderLines?: number;
+  restListenPort?: number;
+  customProperties?: string[];
+  tags?: string[];
+  virtualProxies?: string[];
+}